analyze/requirements.txt
numpy~=2.0.1
pillow~=10.4.0
pip~=23.3
wheel~=0.38.4
pytz~=2024.1
PyYAML~=6.0.2
zipp~=3.20.0
setuptools~=72.2.0
pandas~=2.2.2
matplotlib~=3.9.2
python-dateutil~=2.9.0.post0
certifi~=2024.7.4
seaborn~=0.13.2
cycler~=0.12.1
urllib3~=2.2.2
six~=1.16.0
idna~=3.7
contourpy~=1.2.1
fonttools~=4.53.1
pyparsing~=3.1.2
packaging~=24.1
kiwisolver~=1.4.5
requests~=2.32.3